Make status optional
