--- alembic/versions/2025_01_15_1500-f8419fa049a0_add_status_and_browser_address_to_.py
@@ -1,33 +1,0 @@
-"""add status and browser address to persistent browser session
-
-Revision ID: f8419fa049a0
-Revises: 46e38fc53f64
-Create Date: 2025-01-15 15:00:09.718550+00:00
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "f8419fa049a0"
-down_revision: Union[str, None] = "46e38fc53f64"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("persistent_browser_sessions", sa.Column("browser_address", sa.String(), nullable=True))
-    op.add_column("persistent_browser_sessions", sa.Column("status", sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("persistent_browser_sessions", "status")
-    op.drop_column("persistent_browser_sessions", "browser_address")
-    # ### end Alembic commands ###
