+++ alembic/versions/2025_01_15_1504-9adef4708ca8_persistent_browser_add_status_and_.py
@@ -1,0 +1,32 @@
+"""persistent browser add status and browser address
+
+Revision ID: 9adef4708ca8
+Revises: 46e38fc53f64
+Create Date: 2025-01-15 15:04:38.423921+00:00
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '9adef4708ca8'
+down_revision: Union[str, None] = '46e38fc53f64'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('persistent_browser_sessions', sa.Column('browser_address', sa.String(), nullable=True))
+    op.add_column('persistent_browser_sessions', sa.Column('status', sa.String(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('persistent_browser_sessions', 'status')
+    op.drop_column('persistent_browser_sessions', 'browser_address')
+    # ### end Alembic commands ###
