+++ skyvern/forge/sdk/db/models.py
@@ -592,7 +592,7 @@ class PersistentBrowserSessionModel(Base):
     runnable_id = Column(String, nullable=True)
     browser_id = Column(String, nullable=True)
     browser_address = Column(String, nullable=True)
-    status = Column(String, nullable=False, default="pending")
+    status = Column(String, nullable=True, default="pending")
     created_at = Column(DateTime, default=datetime.datetime.utcnow, nullable=False)
     modified_at = Column(DateTime, default=datetime.datetime.utcnow, onupdate=datetime.datetime.utcnow, nullable=False)
     deleted_at = Column(DateTime, nullable=True)
